Replace login and signup actions with auth
